Add WebAssembly SQL shell with OPFS storage
Provide a browser SQL shell with Memory and OPFS instances, hybrid search and
fork table examples, exact result formatting, and complete runtime cleanup.

Fix wasm32 portability issues and add regression coverage. Batch Worker results,
bound previews and rendered output, reuse compiled modules across new instances,
and package versioned releases with compression and HTTP caching.

+++ src/logservice/common_util/ob_log_time_utils.cpp
@@ -56,7 +56,7 @@ int print_human_tstamp(char *buf, const int64_t buf_len, int64_t &pos,
                                   tm.tm_hour,
                                   tm.tm_min,
                                   tm.tm_sec,
-                                  tv.tv_usec);
+                                  static_cast<int64_t>(tv.tv_usec));
   }
   return ret;
 }
@@ -90,7 +90,7 @@ int print_human_timeval(char *buf,
                                   hr,
                                   min,
                                   sec,
-                                  tv.tv_usec);
+                                  static_cast<int64_t>(tv.tv_usec));
   }
   return ret;
 }
